renamed module from spring-amqp to spring-rabbit (#603)
