+++ instrumentation/spring-rabbit/pom.xml
@@ -8,7 +8,7 @@
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>brave-instrumentation-spring-amqp</artifactId>
+	<artifactId>brave-instrumentation-spring-rabbit</artifactId>
 	<name>Brave Instrumentation: Spring AMQP (RabbitMQ)</name>
 
 	<properties>
